+++ Basestation_Software.Web/Basestation_Software.Web.csproj
@@ -13,7 +13,7 @@
     </PackageReference>
     <PackageReference Include="OpenCvSharp4" Version="4.10.0.20240616" />
     <PackageReference Include="OpenCvSharp4.runtime.ubuntu.22.04-x64" Version="4.6.0-SNAPSHOT" />
-    <PackageReference Include="RoveComm" Version="1.0.2" />
+    <PackageReference Include="RoveComm" Version="1.0.7" />
     <PackageReference Include="ScottPlot" Version="5.0.40" />
     <PackageReference Include="ScottPlot.Blazor" Version="5.0.39" />
     <PackageReference Include="OpenCvSharp4.runtime.win" Version="4.10.0.20240616" />
@@ -34,5 +34,10 @@
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
+
+  <Target Name="PostBuild" AfterTargets="PostBuildEvent">
+    <Exec Command="npm install" WorkingDirectory="Core/JS/" />
+    <Exec Command="npm run build" WorkingDirectory="Core/JS/" />
+  </Target>
 
 </Project>
