+++ Basestation_Software.Web/Core/Components/AutoStateDiagram.razor
@@ -44,7 +44,7 @@
     protected override async Task OnInitializedAsync()
     {
         // Load auto diagram as XElement
-        string diagramFilePath = "wwwroot/state-diagram.svg";
+        string diagramFilePath = "wwwroot/images/state-diagram.svg";
         diagramRoot = XElement.Load(diagramFilePath);
 
         for(int i = 0; i < _stateNames.Length; i++)
