+++ Basestation_Software.Web/Core/Components/RoverMap.razor
@@ -47,7 +47,8 @@
     public int InitialZoomLevel { get; set; } = 18;
 
     // Declare page member variables.
-    private string jsInteropName;
+    private string jsInteropName = "";
+    private IJSObjectReference? jsModule;
 
     // Flags.
     private bool _dataLoaded = false;
@@ -55,7 +56,7 @@
 
     protected override void OnInitialized()
     {
-        this.jsInteropName = $"window.roverMaps.{ID}";
+        this.jsInteropName = $"roverMaps.{ID}";
     }
 
     protected override async Task OnInitializedAsync()
@@ -69,15 +70,16 @@
     {
         if (firstRender)
         {
+            jsModule = await _IJSRuntime.InvokeAsync<IJSObjectReference>("import", "./js/rover-map.js");
             // Create leaflet map.
             Uri baseUri = new Uri(_NavigationManager.BaseUri);
-            await _IJSRuntime.InvokeVoidAsync($"createRoverMap", [
+            await jsModule.InvokeVoidAsync("createRoverMap", [
                 ID,
-                DotNetObjectReference.Create(this),
-                $"{baseUri.Scheme}://{baseUri.Host}:5000/api/MapTiles/{{z}}/{{y}}/{{x}}.png",
-                InitialLat,
-                InitialLong,
-                InitialZoomLevel
+        DotNetObjectReference.Create(this),
+        $"{baseUri.Scheme}://{baseUri.Host}:5000/api/MapTiles/{{z}}/{{y}}/{{x}}.png",
+        InitialLat,
+        InitialLong,
+        InitialZoomLevel
             ]);
 
             // Plot waypoints and keep them updated.
@@ -89,7 +91,7 @@
     private async Task OnWaypointSync()
     {
         // Update waypoint markers.
-        await _IJSRuntime.InvokeVoidAsync($"{jsInteropName}.clearWaypointMarkers");
+        await jsModule!.InvokeVoidAsync($"{jsInteropName}.clearWaypointMarkers");
         await PlotWaypoints();
     }
 
@@ -97,11 +99,11 @@
     {
         await Task.WhenAll(_GPSWaypointService.GetGPSWaypoints().ConvertAll<Task>(waypoint =>
         {
-            return _IJSRuntime.InvokeVoidAsync($"{jsInteropName}.addWaypointMarker", [
+            return jsModule!.InvokeVoidAsync($"{jsInteropName}.addWaypointMarker", [
                 waypoint.Latitude.GetValueOrDefault(0),
-                waypoint.Longitude.GetValueOrDefault(0),
-                waypoint.SearchRadius.GetValueOrDefault(0),
-                ColorTranslator.ToHtml(Color.FromArgb(waypoint.WaypointColor.GetValueOrDefault(0)))
+            waypoint.Longitude.GetValueOrDefault(0),
+            waypoint.SearchRadius.GetValueOrDefault(0),
+            ColorTranslator.ToHtml(Color.FromArgb(waypoint.WaypointColor.GetValueOrDefault(0)))
             ]).AsTask();
         }));
     }
@@ -141,19 +143,21 @@
     {
     for (int x = (tileX - 3); x < (tileX + 3); x++)
     {
-        MapTile? mapTile = await _MapTileService.GetMapTile(x, y, zoom);
-        if (mapTile is null)
-        {
-            string url = $"http://mt1.google.com/vt/lyrs=y&x={x}&y={y}&z={zoom}";
-            tasks.Add(DownloadAndSaveTile(client, url, x, y, zoom));
-            _cachingMapTiles = true;
-        }
+    MapTile? mapTile = await _MapTileService.GetMapTile(x, y, zoom);
+    if (mapTile is null)
+    {
+        string url = $"http://mt1.google.com/vt/lyrs=y&x={x}&y={y}&z={zoom}";
+        tasks.Add(DownloadAndSaveTile(client, url, x, y, zoom));
+        _cachingMapTiles = true;
+    }
     }
     }
 
     await Task.WhenAll(tasks);
     _cachingMapTiles = false;
     }
+#else
+    await Task.CompletedTask;
 #endif
     }
