+++ Basestation_Software.Web/Core/JS/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "basestation-software",
+  "version": "1.0.0",
+  "description": "JS modules used in the MRDT BaseStation software",
+  "author": "MissouriMRDT",
+  "license": "MIT",
+  "scripts": {
+    "build": "snowpack build"
+  },
+  "dependencies": {
+    "leaflet": "^1.9.4",
+    "leaflet-contextmenu": "^1.4.0",
+    "three": "^0.173.0",
+    "three-stdlib": "^2.35.13"
+  },
+  "devDependencies": {
+    "@snowpack/plugin-optimize": "^0.2.13",
+    "snowpack": "^3.8.8"
+  }
+}
